Merge

+++ src/controlesJpa/ClienteJpaController.java
@@ -5,20 +5,24 @@
  */
 package controlesJpa;
 
+import controlesJpa.exceptions.IllegalOrphanException;
 import controlesJpa.exceptions.NonexistentEntityException;
 import entidades.Cliente;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import entidades.Condutor;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
- * @author guilherme.santos
+ * @author luanl
  */
 public class ClienteJpaController implements Serializable {
 
@@ -32,11 +36,29 @@ public class ClienteJpaController implements Serializable {
     }
 
     public void create(Cliente cliente) {
+        if (cliente.getCondutorCollection() == null) {
+            cliente.setCondutorCollection(new ArrayList<Condutor>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Collection<Condutor> attachedCondutorCollection = new ArrayList<Condutor>();
+            for (Condutor condutorCollectionCondutorToAttach : cliente.getCondutorCollection()) {
+                condutorCollectionCondutorToAttach = em.getReference(condutorCollectionCondutorToAttach.getClass(), condutorCollectionCondutorToAttach.getIdCondutor());
+                attachedCondutorCollection.add(condutorCollectionCondutorToAttach);
+            }
+            cliente.setCondutorCollection(attachedCondutorCollection);
             em.persist(cliente);
+            for (Condutor condutorCollectionCondutor : cliente.getCondutorCollection()) {
+                Cliente oldIdClienteOfCondutorCollectionCondutor = condutorCollectionCondutor.getIdCliente();
+                condutorCollectionCondutor.setIdCliente(cliente);
+                condutorCollectionCondutor = em.merge(condutorCollectionCondutor);
+                if (oldIdClienteOfCondutorCollectionCondutor != null) {
+                    oldIdClienteOfCondutorCollectionCondutor.getCondutorCollection().remove(condutorCollectionCondutor);
+                    oldIdClienteOfCondutorCollectionCondutor = em.merge(oldIdClienteOfCondutorCollectionCondutor);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -45,12 +67,45 @@ public class ClienteJpaController implements Serializable {
         }
     }
 
-    public void edit(Cliente cliente) throws NonexistentEntityException, Exception {
+    public void edit(Cliente cliente) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Cliente persistentCliente = em.find(Cliente.class, cliente.getIdCliente());
+            Collection<Condutor> condutorCollectionOld = persistentCliente.getCondutorCollection();
+            Collection<Condutor> condutorCollectionNew = cliente.getCondutorCollection();
+            List<String> illegalOrphanMessages = null;
+            for (Condutor condutorCollectionOldCondutor : condutorCollectionOld) {
+                if (!condutorCollectionNew.contains(condutorCollectionOldCondutor)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Condutor " + condutorCollectionOldCondutor + " since its idCliente field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Collection<Condutor> attachedCondutorCollectionNew = new ArrayList<Condutor>();
+            for (Condutor condutorCollectionNewCondutorToAttach : condutorCollectionNew) {
+                condutorCollectionNewCondutorToAttach = em.getReference(condutorCollectionNewCondutorToAttach.getClass(), condutorCollectionNewCondutorToAttach.getIdCondutor());
+                attachedCondutorCollectionNew.add(condutorCollectionNewCondutorToAttach);
+            }
+            condutorCollectionNew = attachedCondutorCollectionNew;
+            cliente.setCondutorCollection(condutorCollectionNew);
             cliente = em.merge(cliente);
+            for (Condutor condutorCollectionNewCondutor : condutorCollectionNew) {
+                if (!condutorCollectionOld.contains(condutorCollectionNewCondutor)) {
+                    Cliente oldIdClienteOfCondutorCollectionNewCondutor = condutorCollectionNewCondutor.getIdCliente();
+                    condutorCollectionNewCondutor.setIdCliente(cliente);
+                    condutorCollectionNewCondutor = em.merge(condutorCollectionNewCondutor);
+                    if (oldIdClienteOfCondutorCollectionNewCondutor != null && !oldIdClienteOfCondutorCollectionNewCondutor.equals(cliente)) {
+                        oldIdClienteOfCondutorCollectionNewCondutor.getCondutorCollection().remove(condutorCollectionNewCondutor);
+                        oldIdClienteOfCondutorCollectionNewCondutor = em.merge(oldIdClienteOfCondutorCollectionNewCondutor);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -68,7 +123,7 @@ public class ClienteJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -79,6 +134,17 @@ public class ClienteJpaController implements Serializable {
                 cliente.getIdCliente();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The cliente with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            Collection<Condutor> condutorCollectionOrphanCheck = cliente.getCondutorCollection();
+            for (Condutor condutorCollectionOrphanCheckCondutor : condutorCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Cliente (" + cliente + ") cannot be destroyed since the Condutor " + condutorCollectionOrphanCheckCondutor + " in its condutorCollection field has a non-nullable idCliente field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(cliente);
             em.getTransaction().commit();
